requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cssselect==1.2.0
cssutils==2.9.0
Django==5.0
django-axes==6.2.0
django-crispy-forms==2.1
gunicorn==21.2.0
idna==3.6
lxml==4.9.4
packaging==23.2
premailer==3.10.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
yagmail==0.15.293
